backend/requirements.txt
aiofiles==23.2.1
fastapi==0.115.0
uvicorn==0.31.0
chromadb==0.5.12
sentence-transformers==3.1.1
pydub==0.25.1
gTTS==2.2.4
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
